+++ matdb/utility.py
@@ -913,11 +913,10 @@ def check_deps():
             name, version = pkg, None
         if name in req_pckgs:
             count = version.count(".")
-            if (version is not None and (version.count(".") == 2 or
-                                         version.count(".") == 3) and ("/" not in version
-                                                                       and ":" not in version
-                                                                       and  "-" not in version
-                                                                       and ".com" not in version)):
+            if version is not None and (count in [1,2,3] and ("/" not in version
+                                                              and ":" not in version
+                                                              and  "-" not in version
+                                                              and ".com" not in version)):
                 versions[name] = version
                 req_pckgs.remove(name)
             else: #pragma: no cover There won't be locally installed
@@ -935,7 +934,7 @@ def required_packages():
     hard coded before each commit.
     """
 
-    return ["ase", "backports.functools-lru-cache", "beautifulsoup4", "certifi",
+    return ["ase", "beautifulsoup4", "certifi",
             "chardet", "cycler", "h5py", "html5lib", "idna", "matplotlib", "mpld3",
             "numpy", "phenum", "phonopy", "pyparsing", "python-dateutil", "pytz",
             "PyYAML", "requests", "six", "subprocess32", "termcolor",